U26-42 Добавить API для лидерборда
