+++ packages/client/src/components/Game/Game.tsx
@@ -2,7 +2,6 @@ import React, { useEffect, useRef, useState } from 'react'
 import GameEngine from '../../domain/Game/GameEngine'
 import classes from './Game.module.scss'
 import { leaderboardController } from '../../controllers/LeaderboardController'
-import { useAppSelector } from '../../store/hooks'
 
 enum GameStates {
   GAME_BEFORE_START,
@@ -16,24 +15,8 @@ const Game: React.FC = () => {
   const gameDiv = useRef<HTMLDivElement>(null)
   const gameEngineRef = useRef<GameEngine>()
 
-  const user = useAppSelector(state => state.user)
-
   const endGameCallback = (points: number) => {
-    const now = new Date()
-    const month = now.getMonth() + 1
-    const date =
-      now.getDate() +
-      '.' +
-      (month < 10 ? '0' + month : month) +
-      '.' +
-      now.getFullYear()
-    leaderboardController.addToLeaderboard({
-      data: {
-        codeNinjasScore: points,
-        login: user?.login,
-        date,
-      },
-    })
+    leaderboardController.addToLeaderboard(points)
     setScore(points)
     setGameState(GameStates.GAME_END)
   }
